+++ core/domain/src/main/java/com/example/domain/user/UserRepository.kt
@@ -37,7 +37,7 @@ interface UserRepository {
         username: String,
         user_pass: String,
         email: String,
-        mobile_number: Int,
+        mobile_number: Long,
         gender: String,
         address: String,
         profile_pic: String? = null
