+++ feature/authentication/src/main/java/com/example/authentication/register/RegisterEvent.kt
@@ -1,6 +1,7 @@
 package com.example.authentication.register
 
 import com.example.authentication.login.LoginEvent
+import java.math.BigInteger
 
 sealed class RegisterEvent{
     class OnFirstNameChange(val first_name: String): RegisterEvent()
@@ -9,7 +10,7 @@ sealed class RegisterEvent{
     class OnUserPassChange(val user_pass: String): RegisterEvent()
     class OnConfirmUserPassChange(val confirmUserPass: String): RegisterEvent()
     class OnGenderChange(val gender: String): RegisterEvent()
-    class OnMobileNumberChange(val mobile_number: Int): RegisterEvent()
+    class OnMobileNumberChange(val mobile_number: Long): RegisterEvent()
     class OnEmailChange(val email: String): RegisterEvent()
     class OnAddressChange(val address: String): RegisterEvent()
     class OnProfilePicChange(val profile_pic: String): RegisterEvent()
